requirements.txt
absl-py==0.2.2
alabaster==0.7.11
appnope==0.1.0
arrow==0.12.1
astor==0.7.1
astroid==2.1.0
atomicwrites==1.1.5
attrs==18.1.0
autograd==1.2
Babel==2.6.0
backcall==0.1.0
binaryornot==0.4.4
bleach==1.5.0
CausalInference==0.1.2
certifi==2018.4.16
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
cookiecutter==1.6.0
coverage==4.5.1
cycler==0.10.0
dask==0.20.1
decorator==4.3.0
defusedxml==0.5.0
docutils==0.14
entrypoints==0.2.3
fix-yahoo-finance==0.0.22
flake8==3.6.0
future==0.16.0
gast==0.2.0
grpcio==1.13.0
h5py==2.8.0
html5lib==0.9999999
idna==2.7
imagesize==1.0.0
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.12.1
Jinja2==2.10
jinja2-time==0.2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter-dashboards==0.7.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lifelines==0.14.6
lightgbm==2.2.2
lime==0.1.1.32
lxml==4.2.5
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.2
mccabe==0.6.1
memory-profiler==0.55.0
mistune==0.8.4
more-itertools==4.2.0
multitasking==0.0.7
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.0
numpy==1.15.4
opencv-python==3.4.3.18
packaging==17.1
pandas==0.23.4
pandas-datareader==0.7.0
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.3.0
pluggy==0.6.0
poyo==0.4.2
prometheus-client==0.4.1
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.6.2
psycopg2==2.7.5
ptyprocess==0.6.0
py==1.5.4
pycodestyle==2.4.0
pydicom==1.2.2
pyflakes==2.0.0
Pygments==2.2.0
pylint==2.2.2
pyparsing==2.2.0
pytest==3.6.3
pytest-cov==2.5.1
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
scikit-image==0.14.1
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.7.6
sphinx-rtd-theme==0.4.0
sphinxcontrib-websupport==1.1.0
statsmodels==0.9.0
TA-Lib==0.4.17
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.3
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
typed-ast==1.3.1
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
whichcraft==0.5.2
widgetsnbextension==3.4.2
wordcloud==1.5.0
wrapt==1.10.11
xgboost==0.81
